Resolved Pitch Encoder Issues
I tried to do this as cleanly as possible, but it's a little messy. In the profile, 'knob position' actually means *VALUE*, where the actual knob position varies depending on the 'mode'.

This code mostly moves the calculations across to the profile, with the daemon only setting up and rechecking when appropriate.

+++ usb/src/goxlr.rs
@@ -375,9 +375,9 @@ impl<T: UsbContext> GoXLR<T> {
     pub fn set_encoder_value(
         &mut self,
         encoder: EncoderName,
-        value: u8,
+        value: i8,
     ) -> Result<(), rusb::Error> {
-        self.request_data(Command::SetEncoderValue(encoder), &[value])?;
+        self.request_data(Command::SetEncoderValue(encoder), &[value as u8])?;
         Ok(())
     }
